wxNviz: point size and height are float (fix #2072)
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@57624 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ gui/wxpython/core/workspace.py
@@ -370,11 +370,11 @@ class ProcessWorkspaceFile:
                                             nviz['vector']['points'])
                 
                 # height
-                self.__processLayerNvizNode(node_vpoints, 'height', int,
+                self.__processLayerNvizNode(node_vpoints, 'height', float,
                                             nviz['vector']['points'])
                 
                 # height
-                self.__processLayerNvizNode(node_vpoints, 'size', int,
+                self.__processLayerNvizNode(node_vpoints, 'size', float,
                                             nviz['vector']['points'])
                 
                 # thematic
